Do not use forms for simple navigation
commandButton and commandLink must be wrapped in a form, since they
result in a POST request. However, this is not needed for simple
navigation in which we simply want to go to another page. In that
case, we can use button and link, which result in a GET request,
and which do not need to be wrapped in a form.

NOTE: the outcome of button and link is evaluated when the view is
rendered. So if an outcome is set to logOut(), then logOut() will
be executed when the page is loaded. In these cases commandButton
should be used instead.